refactor: moved to create user directory
