+++ app/src/main/java/jp/aoyama/mki/thermometer/fragment/user/create/BluetoothPairingFragment.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.fragment.user
+package jp.aoyama.mki.thermometer.fragment.user.create
 
 import android.content.Intent
 import android.os.Bundle
