+++ app/src/main/java/jp/aoyama/mki/thermometer/fragment/user/create/SelectBluetoothDeviceFragment.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.fragment.user
+package jp.aoyama.mki.thermometer.fragment.user.create
 
 import android.Manifest
 import android.bluetooth.BluetoothAdapter
